requirements.txt
torch
torchvision
tensorflow
numpy
opencv-python
pillow
scikit-learn
matplotlib
requests
beautifulsoup4
